+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Deprecated the `AltResponseCodes` concept in favor of using all codes 300-599 as "blocking"
+* Added HTTP redirect support
 
 ## 1.7.1 2020-04-06